Move utilities from ParseUtil into Handler, rename to canonicalizeString, call file.length() only once.
