+++ src/java.base/share/classes/sun/net/www/ParseUtil.java
@@ -225,51 +225,6 @@ public final class ParseUtil {
         return sb.toString();
     }
 
-    /**
-     * Returns a canonical version of the specified string.
-     */
-    public static String canonizeString(String file, int bangSlash) {
-        int len = file.length()-bangSlash;
-        if (len == 0 || (file.indexOf("./", bangSlash) == -1 && file.charAt(file.length() - 1) != '.')) {
-            return file;
-        } else {
-            String before = file.substring(0, bangSlash);
-            String after = file.substring(bangSlash);
-            return before + doCanonize(after);
-        }
-    }
-
-    private static String doCanonize(String file) {
-        int i, lim;
-
-        // Remove embedded /../
-        while ((i = file.indexOf("/../")) >= 0) {
-            if ((lim = file.lastIndexOf('/', i - 1)) >= 0) {
-                file = file.substring(0, lim) + file.substring(i + 3);
-            } else {
-                file = file.substring(i + 3);
-            }
-        }
-        // Remove embedded /./
-        while ((i = file.indexOf("/./")) >= 0) {
-            file = file.substring(0, i) + file.substring(i + 2);
-        }
-        // Remove trailing ..
-        while (file.endsWith("/..")) {
-            i = file.indexOf("/..");
-            if ((lim = file.lastIndexOf('/', i - 1)) >= 0) {
-                file = file.substring(0, lim+1);
-            } else {
-                file = file.substring(0, i);
-            }
-        }
-        // Remove trailing .
-        if (file.endsWith("/."))
-            file = file.substring(0, file.length() -1);
-
-        return file;
-    }
-
     public static URL fileToEncodedURL(File file)
         throws MalformedURLException
     {
