+++ src/java.base/share/classes/sun/net/www/protocol/jar/Handler.java
@@ -27,7 +27,6 @@ package sun.net.www.protocol.jar;
 
 import java.io.IOException;
 import java.net.*;
-import sun.net.www.ParseUtil;
 
 /*
  * Jar URL Handler
@@ -166,7 +165,7 @@ public class Handler extends java.net.URLStreamHandler {
 
             // Canonize the result after the bangslash
             int bangSlash = indexOfBangSlash(file);
-            file = ParseUtil.canonizeString(file, bangSlash);
+            file = canonicalizeString(file, bangSlash);
         }
         setURL(url, "jar", "", -1, file, ref);
     }
@@ -212,5 +211,52 @@ public class Handler extends java.net.URLStreamHandler {
             }
         }
         return (ctxFile + spec);
+    }
+
+    /**
+     * Returns a version of the specified string with
+     * canonicalization applied starting from position {@code off}
+     */
+    private static String canonicalizeString(String file, int off) {
+        int len = file.length();
+        if (off >= len || (file.indexOf("./", off) == -1 && file.charAt(len - 1) != '.')) {
+            return file;
+        } else {
+            // Defer substring and concat until canonicalization is required
+            String before = file.substring(0, off);
+            String after = file.substring(off);
+            return before + doCanonize(after);
+        }
+    }
+
+    private static String doCanonize(String file) {
+        int i, lim;
+
+        // Remove embedded /../
+        while ((i = file.indexOf("/../")) >= 0) {
+            if ((lim = file.lastIndexOf('/', i - 1)) >= 0) {
+                file = file.substring(0, lim) + file.substring(i + 3);
+            } else {
+                file = file.substring(i + 3);
+            }
+        }
+        // Remove embedded /./
+        while ((i = file.indexOf("/./")) >= 0) {
+            file = file.substring(0, i) + file.substring(i + 2);
+        }
+        // Remove trailing ..
+        while (file.endsWith("/..")) {
+            i = file.indexOf("/..");
+            if ((lim = file.lastIndexOf('/', i - 1)) >= 0) {
+                file = file.substring(0, lim+1);
+            } else {
+                file = file.substring(0, i);
+            }
+        }
+        // Remove trailing .
+        if (file.endsWith("/."))
+            file = file.substring(0, file.length() -1);
+
+        return file;
     }
 }
